fixed:
https://github.com/Oljk/base_new_app_1/pull/3#discussion_r558844816
https://github.com/Oljk/base_new_app_1/pull/3#discussion_r558844248
https://github.com/Oljk/base_new_app_1/pull/3#discussion_r558844503
