+++ src/components/header/Logo.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Logo(props) {
+    return (<img img={props.logoUrl} className="App-logo" alt={"logo"}/>);
+}
+
+export {Logo};
